update tarkov tracker on every page
